# internal/gateway/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: SirenaIX Messaging Gateway API
  version: 1.0.0
  description: |
    Tenant-scoped connection, pairing, and contact metadata operations. Deploy
    behind HTTPS. Google Messages pairing uses the official app on a physical,
    online Android phone; it needs no custom APK or continuously open laptop.
    Encrypted sessions survive ordinary gateway restarts, but Google may revoke
    them and require the explicit reauthorization flow.
servers:
  - url: /
security:
  - bearerAuth: []
paths:
  /v1/connections:
    get:
      operationId: listConnections
      summary: List the authenticated tenant's connections
      parameters:
        - {name: after, in: query, schema: {type: string, maxLength: 256}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
      responses:
        '200':
          description: Connections
          content:
            application/json:
              schema:
                type: object
                required: [connections]
                properties:
                  connections:
                    type: array
                    items: {$ref: '#/components/schemas/Connection'}
                  next_cursor: {type: string}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '413': {$ref: '#/components/responses/RequestTooLarge'}
    post:
      operationId: createConnection
      summary: Create an unpaired Google Messages physical-phone connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [display_label]
              properties:
                display_label: {type: string, minLength: 1, maxLength: 128}
      responses:
        '201': {description: Unpaired connection created, content: {application/json: {schema: {$ref: '#/components/schemas/Connection'}}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '413': {$ref: '#/components/responses/RequestTooLarge'}
        '415': {$ref: '#/components/responses/UnsupportedMediaType'}
...
  /v1/connections/{connection_id}/messages:
    parameters:
      - $ref: '#/components/parameters/ConnectionID'
    post:
      operationId: submitMessage
      summary: Queue one idempotent outbound message
      description: |
        Existing conversations and groups use the phone-selected transport and
        default outgoing SIM route. New chats are one-to-one E.164 only and
        require route_mode=phone_default with line_id omitted. Relay acceptance
        is provider_accepted, not carrier delivery. Because Google exposes no
        provider idempotency guarantee, an ambiguous dispatch is uncertain and
        is never blindly retried.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SendMessageRequest'}
      responses:
        '202': {description: Message durably queued, content: {application/json: {schema: {$ref: '#/components/schemas/Message'}}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '409': {$ref: '#/components/responses/Conflict'}
        '413': {$ref: '#/components/responses/RequestTooLarge'}
        '415': {$ref: '#/components/responses/UnsupportedMediaType'}
        '500': {$ref: '#/components/responses/InternalError'}
  /v1/messages:
    get:
      operationId: listMessages
      parameters:
        - {name: after, in: query, required: false, schema: {type: string}}
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
      responses:
        '200': {description: Bounded message page, content: {application/json: {schema: {$ref: '#/components/schemas/MessagePage'}}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '413': {$ref: '#/components/responses/RequestTooLarge'}
        '500': {$ref: '#/components/responses/InternalError'}
  /v1/messages/{message_id}:
    parameters:
      - $ref: '#/components/parameters/MessageID'
    get:
      operationId: getMessage
      responses:
        '200': {description: Message and conservative current status, content: {application/json: {schema: {$ref: '#/components/schemas/Message'}}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '413': {$ref: '#/components/responses/RequestTooLarge'}
        '500': {$ref: '#/components/responses/InternalError'}
  /v1/media:
    post:
      operationId: uploadMedia
      summary: Upload one bounded passive raster image
      description: JPEG, PNG, GIF, or WebP only; SVG and active content are rejected.
      requestBody:
        required: true
        content:
          image/jpeg: {schema: {type: string, format: binary}}
          image/png: {schema: {type: string, format: binary}}
          image/gif: {schema: {type: string, format: binary}}
          image/webp: {schema: {type: string, format: binary}}
      responses:
        '201': {description: Media stored, content: {application/json: {schema: {$ref: '#/components/schemas/Media'}}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '413': {$ref: '#/components/responses/RequestTooLarge'}
        '415': {$ref: '#/components/responses/UnsupportedMediaType'}
        '500': {$ref: '#/components/responses/InternalError'}
...
  /v1/webhooks:
    get:
      operationId: listWebhooks
      parameters:
        - {name: after, in: query, schema: {type: string, maxLength: 256}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
      responses:
        '200': {description: Webhook endpoints without secrets, content: {application/json: {schema: {type: object, properties: {webhooks: {type: array, items: {$ref: '#/components/schemas/WebhookEndpoint'}}, next_cursor: {type: string}}}}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '413': {$ref: '#/components/responses/RequestTooLarge'}
        '500': {$ref: '#/components/responses/InternalError'}
    post:
      operationId: createWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, additionalProperties: false, required: [destination], properties: {destination: {type: string, format: uri}}}
      responses:
        '201': {description: Endpoint and one-time secret, content: {application/json: {schema: {$ref: '#/components/schemas/WebhookCreatedResponse'}}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '413': {$ref: '#/components/responses/RequestTooLarge'}
        '415': {$ref: '#/components/responses/UnsupportedMediaType'}
        '500': {$ref: '#/components/responses/InternalError'}
  /v1/webhooks/{webhook_id}:
    parameters:
      - $ref: '#/components/parameters/WebhookID'
    delete:
      operationId: deleteWebhook
      responses:
        '202':
          description: Endpoint disabled; an already-started request is still draining
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [status]
                properties:
                  status: {type: string, enum: [deleting]}
        '204': {description: Endpoint disabled without deleting history}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '413': {$ref: '#/components/responses/RequestTooLarge'}
        '500': {$ref: '#/components/responses/InternalError'}
...
  /v1/webhooks/dlq/{dlq_id}:replay:
    parameters:
      - $ref: '#/components/parameters/DLQID'
    post:
      operationId: replayWebhookDLQ
      responses:
        '202': {description: Delivery retry cycle reset, content: {application/json: {schema: {type: object}}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '413': {$ref: '#/components/responses/RequestTooLarge'}
        '500': {$ref: '#/components/responses/InternalError'}
  /v1/contacts:
    get:
      operationId: listContacts
      summary: List contacts with server metadata
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: cursor
          in: query
          required: false
          description: Opaque cursor returned by the preceding page.
          schema: {type: string, maxLength: 683}
      responses:
        '200':
          description: Contact page
          content:
            application/json:
              schema:
                type: object
                required: [contacts, next_cursor]
                properties:
                  contacts:
                    type: array
                    items: {$ref: '#/components/schemas/Contact'}
                  next_cursor:
                    type: string
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '413': {$ref: '#/components/responses/RequestTooLarge'}
    put:
      operationId: upsertServerContact
      summary: Idempotently create or update a tenant-local contact
      description: Creates a contact for a canonical E.164 number before provider import, without writing any data back to the phone. Omitting server_alias preserves its current value; null or an empty string clears it.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ServerContactUpsert'}
      responses:
        '200': {description: Tenant-local contact converged by phone number, content: {application/json: {schema: {$ref: '#/components/schemas/Contact'}}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '413': {$ref: '#/components/responses/RequestTooLarge'}
        '415': {$ref: '#/components/responses/UnsupportedMediaType'}
        '500': {$ref: '#/components/responses/InternalError'}
  /v1/contacts/{contact_id}:
    parameters:
      - $ref: '#/components/parameters/ContactID'
    patch:
      operationId: patchContact
      summary: Set or clear the server-owned alias
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [server_alias]
              properties:
                server_alias:
                  type: string
                  nullable: true
      responses:
        '204': {description: Alias updated}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '413': {$ref: '#/components/responses/RequestTooLarge'}
        '415': {$ref: '#/components/responses/UnsupportedMediaType'}
  /v1/labels:
    get:
      operationId: listLabels
      summary: List tenant-owned labels
      responses:
        '200':
          description: Labels
          content:
            application/json:
              schema:
                type: object
                required: [labels]
                properties:
                  labels:
                    type: array
                    items: {$ref: '#/components/schemas/Label'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '413': {$ref: '#/components/responses/RequestTooLarge'}
    post:
      operationId: createLabel
      summary: Create or converge on a normalized tenant label
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name: {type: string}
      responses:
        '200': {description: Existing normalized label, content: {application/json: {schema: {$ref: '#/components/schemas/Label'}}}}
        '201': {description: Label created, content: {application/json: {schema: {$ref: '#/components/schemas/Label'}}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '413': {$ref: '#/components/responses/RequestTooLarge'}
        '415': {$ref: '#/components/responses/UnsupportedMediaType'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ConnectionID:
      name: connection_id
      in: path
      required: true
      schema: {type: string}
    ContactID:
      name: contact_id
      in: path
      required: true
      schema: {type: string}
...
    MessageID:
      name: message_id
      in: path
      required: true
      schema: {type: string}
...
    WebhookID:
      name: webhook_id
      in: path
      required: true
      schema: {type: string}
    DLQID:
      name: dlq_id
      in: path
      required: true
      schema: {type: string}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 200}
  responses:
    InvalidRequest:
      description: Invalid request
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
    Unauthorized:
      description: Missing or invalid bearer token
      headers:
        WWW-Authenticate: {schema: {type: string}}
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
    Forbidden:
      description: Required scope missing
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
    NotFound:
      description: Tenant-owned object was not found
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
    MethodNotAllowed:
      description: Method is not allowed for the resource
      headers:
        Allow: {schema: {type: string}}
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
    RequestTooLarge:
      description: Request body exceeds the supported size
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
    UnsupportedMediaType:
      description: Request body must use application/json
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
    Conflict:
      description: Pairing state conflict
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
...
    InternalError:
      description: Pairing or persistence operation failed safely
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
  schemas:
    Connection:
      type: object
      required: [id, name, state]
      properties:
        id: {type: string}
        name: {type: string}
        state: {$ref: '#/components/schemas/ConnectionState'}
    SendMessageRequest:
      type: object
      additionalProperties: false
      properties:
        conversation_id: {type: string}
        recipient: {type: string, pattern: '^\+[1-9][0-9]{7,14}$'}
        line_id: {type: string}
        route_mode: {type: string, enum: [phone_default]}
        text: {type: string, maxLength: 65536}
        media_ids: {type: array, maxItems: 16, uniqueItems: true, items: {type: string}}
    Message:
      type: object
      required: [id, connection_id, direction, state, text, created_at]
      properties:
        id: {type: string}
        connection_id: {type: string}
        conversation_id: {type: string, description: Present after an existing conversation is selected or a new chat is resolved by the phone.}
        direction: {type: string, enum: [inbound, outbound, unknown]}
        provider_message_id: {type: string, description: Provider identity when observed; absent before remote evidence.}
        transport:
          type: string
          enum: [sms, mms, rcs]
          description: Observed provider transport only. Callers cannot select or force SMS, MMS, or RCS.
        recipient: {type: string}
        line_id: {type: string}
        route_mode: {type: string}
        text: {type: string}
        media_ids: {type: array, items: {type: string}}
        attachments:
          type: array
          maxItems: 16
          items: {$ref: '#/components/schemas/MessageAttachment'}
        state: {$ref: '#/components/schemas/MessageState'}
        created_at: {type: string, format: date-time}
...
    MessagePage:
      type: object
      required: [messages]
      properties:
        messages: {type: array, items: {$ref: '#/components/schemas/Message'}}
        next_cursor: {type: string}
    Media:
      type: object
      description: Inbound media is visible as pending before bytes and decoded dimensions are available.
      required: [id, state, display_filename, created_at]
      properties:
        id: {type: string}
        state: {type: string, enum: [pending, ready, failed]}
        mime_type: {type: string, enum: [image/jpeg, image/png, image/gif, image/webp]}
        size: {type: integer, format: int64, maximum: 26214400}
        width: {type: integer, minimum: 1}
        height: {type: integer, minimum: 1}
        display_filename: {type: string}
        created_at: {type: string, format: date-time}
        content_path: {type: string}
    WebhookEndpoint:
      type: object
      required: [id, destination, key_id, active, created_at]
      properties:
        id: {type: string}
        destination: {type: string, format: uri}
        key_id: {type: string}
        active: {type: boolean}
        created_at: {type: string, format: date-time}
...